# services/api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moventra API Foundation",
    "version": "0.1.0",
    "description": "Personal and enterprise read boundaries plus business upgrade intent. No production money operations. All API responses are no-store."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin API; local server defaults to port 8870. No production deployment yet."
    }
  ],
  "security": [
    {
      "firebaseBearer": []
    }
  ],
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "getIdentity",
        "responses": {
          "200": {
            "description": "Authenticated identity and owned/member subjects",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "customers": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Customer"
                          }
                        },
                        "operator": {
                          "type": "boolean",
                          "description": "Has one or more explicit staff grants; not a global authorization."
                        },
                        "mfaVerified": {
                          "type": "boolean"
                        },
                        "requiresMfa": {
                          "type": "boolean"
                        },
                        "staffScopes": {
                          "type": "array",
                          "description": "Empty until MFA is verified. Every query rechecks authorization.",
                          "items": {
                            "type": "object",
                            "required": [
                              "customerId",
                              "name",
                              "permission"
                            ],
                            "properties": {
                              "customerId": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "name": {
                                "type": "string"
                              },
                              "permission": {
                                "type": "string",
                                "enum": [
                                  "accounts:read",
                                  "transactions:read"
                                ]
                              }
                            }
                          }
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "customer",
                            "admin"
                          ]
                        },
                        "globalAdmin": {
                          "type": "boolean",
                          "description": "Active global administrator; true only after verified operator MFA. Resource authorization and execution gates still apply."
                        }
                      },
                      "required": [
                        "id",
                        "customers",
                        "operator",
                        "mfaVerified",
                        "requiresMfa",
                        "staffScopes",
                        "role"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, revoked or unverified-email identity",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "registration_required when verified UID has no local user; user_disabled when local user exists but is inactive. Only registration_required should offer registration.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency or audit unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "summary": "Read identity and accessible customer subjects"
      }
    },
...
    "/client-api/v1/customers/{customerID}/business-upgrade": {
      "get": {
        "operationId": "getLatestBusinessUpgrade",
        "parameters": [
          {
            "name": "customerID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Requested subject, verified against server-side ownership or grants."
          }
        ],
        "responses": {
          "200": {
            "description": "Upgrade application; submission does not create or activate business accounts",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Upgrade"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, revoked or unverified-email identity",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Local user not active or staff MFA absent",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or inaccessible subject",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency or audit unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "summary": "Read the latest personal-to-business upgrade application"
      },
      "post": {
        "operationId": "submitBusinessUpgrade",
        "description": "Personal owner only. An intent application, not complete KYB. Replays with the same key and normalized name return 200. First submission returns 201. Different payload or another open application returns 409. No financial or ownership mutation.",
        "parameters": [
          {
            "name": "customerID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Requested subject, verified against server-side ownership or grants."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legalName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                },
                "required": [
                  "legalName"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upgrade application; submission does not create or activate business accounts",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Upgrade"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "201": {
            "description": "Upgrade application; submission does not create or activate business accounts",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Upgrade"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, revoked or unverified-email identity",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Local user not active or staff MFA absent",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or inaccessible subject",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Upgrade or idempotency conflict",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "application/json required",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency or audit unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "summary": "Submit a personal-to-business upgrade intent"
      }
    },
    "/api/v1/register": {
      "post": {
        "operationId": "registerUser",
        "summary": "Create the verified Firebase UID login user only; no business subjects or grants",
        "description": "Verified email and non-revoked Firebase token required. Passwords go only to Firebase. Idempotent by Firebase UID; existing users are not modified or reactivated.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created or already active user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "invalid_registration; strict JSON body up to 4096 bytes"
          },
          "401": {
            "description": "unauthenticated"
          },
          "403": {
            "description": "user_disabled; never reactivates an account"
          },
          "415": {
            "description": "invalid_content_type"
          },
          "503": {
            "description": "temporarily_unavailable"
          }
        }
      }
    },
...
    "/admin-api/v1/channel-projections": {
      "get": {
        "operationId": "listChannelProjections",
        "summary": "MFA and explicit channel grant required; sanitized manual projection, never a ledger",
        "parameters": [],
        "responses": {
          "200": {
            "description": "data contains authorized connections, or rows/total/page/revision/sourceAt/importedAt/complete=false/syncMode=manual_import/coverageReason. Admin card and transaction rows include assignmentKind=project_wallet|test_snapshot|unassigned and internal={ownershipStatus:bound|unassigned|scope_mismatch,customerId:UUID|null,userId:UUID|null,customerName:string|null}. Identity is returned only for matching assignment scope; transactions describe current card assignment, not historical funds ownership. No email or Firebase UID. Client DTOs omit these fields"
          },
          "400": {
            "description": "Invalid or unknown query parameter"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "MFA or channel grant missing"
          },
          "404": {
            "description": "Not found or outside granted connection"
          },
          "409": {
            "description": "Projection revision changed"
          },
          "503": {
            "description": "Database or mandatory audit unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "firebaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token",
        "description": "Verified email, valid Firebase identity and active locally provisioned user required. Admin reads additionally require MFA and explicit DB grants."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              }
            },
            "required": [
              "code"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "personal",
              "business"
            ]
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "name"
        ],
        "additionalProperties": false
      },
...
      "Upgrade": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "personalCustomerId": {
            "type": "string",
            "format": "uuid"
          },
          "legalName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "submitted",
              "approved",
              "rejected"
            ]
          },
          "businessCustomerId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "personalCustomerId",
          "legalName",
          "status",
          "businessCustomerId",
          "createdAt"
        ],
        "additionalProperties": false
      },
...
    }
  }
}